+++ allauth/socialaccount/providers/google/views.py
@@ -1,5 +1,7 @@
 import requests
 
+import jwt
+
 from allauth.socialaccount.providers.oauth2.views import (
     OAuth2Adapter,
     OAuth2CallbackView,
@@ -13,16 +15,17 @@ class GoogleOAuth2Adapter(OAuth2Adapter):
     provider_id = GoogleProvider.id
     access_token_url = "https://oauth2.googleapis.com/token"
     authorize_url = "https://accounts.google.com/o/oauth2/v2/auth"
-    profile_url = "https://openidconnect.googleapis.com/v1/userinfo"
 
-    def complete_login(self, request, app, token, **kwargs):
-        resp = requests.get(
-            self.profile_url,
-            params={"access_token": token.token, "alt": "json"},
+    def complete_login(self, request, app, token, response, **kwargs):
+        identity_data = jwt.decode(
+            response["id_token"],
+            # We are sure that the token was received in a response from
+            # Google, so we skip requesting keys for verifying the
+            # signature.
+            options={"verify_signature": False},
+            audience=app.client_id,
         )
-        resp.raise_for_status()
-        extra_data = resp.json()
-        login = self.get_provider().sociallogin_from_response(request, extra_data)
+        login = self.get_provider().sociallogin_from_response(request, identity_data)
         return login